ai_fitness/requirements.txt
streamlit
openai
python-dotenv
pandas
numpy
requests
google-generativeai
groq